requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
et-xmlfile==1.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
openapi-codec==1.3.2
openpyxl==3.1.2
pandas==1.5.3
Pillow==9.4.0
pyaes==1.6.1
pyasn1==0.4.8
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
rsa==4.9
simplejson==3.18.3
six==1.16.0
sydney-py
sqlparse==0.4.3
Telethon==1.27.0
tzdata==2022.7
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.14